% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_u5_env.R
\name{agg_u5_env}
\alias{agg_u5_env}
\title{Aggregrate Under-5 Age Groups in the Under-5 Envelope}
\usage{
agg_u5_env(u5_env, gbd_year)
}
\arguments{
\item{u5_env}{data.table with variables sim, ihme_loc_id, year, sex, age, and deaths}

\item{gbd_year}{}
}
\value{
data.table with variables sim, ihme_loc_id, year, sex, age, and deaths
}
\description{
Aggregate the under-5 envelope from granular age groups (1, 2, 3, 4) to gbd age groups (1, 2-4)
}
\examples{

}
