% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_95_no_hmd_mx.R
\name{calc_95_no_hmd_mx}
\alias{calc_95_no_hmd_mx}
\title{Replace 95+ death rate for envelope, in non-HMD countries, with fitted relationship between 90-94 and 95+ derived from HMD}
\usage{
calc_95_no_hmd_mx(env_mx, mx_params, id_vars)
}
\arguments{
\item{env_mx}{data.table with variables: ihme_loc_id, sex, age, year, sim, mx}

\item{mx_params}{data.table with variables: sex, parmx. Regression parameters based on HMD.}

\item{id_vars}{character vector of id variables (last one must be age). Must uniquely identify observations.}
}
\value{
data.table with all existing variables from env_mx. Modified mx value.
}
\description{
Given a data.table with mx for age 90 and HMD regression parameters, generate mx for age 95.
}
