% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mahalanobis.R
\name{calc_mahalanobis}
\alias{calc_mahalanobis}
\title{Calculate Mahalanobis Function}
\usage{
calc_mahalanobis(draw_num, lt_draws, empir_lt)
}
\arguments{
\item{draw_num}{numeric, row number of the lt_draws data.table to use}

\item{lt_draws}{data.table with variables: ihme_loc_id, year, logit_calc_45q15, logit_calc_5q0
Observations of the HIV-free results from the 5q0 and 45q15 processes (location/age/sex/year/draw-specific)}

\item{empir_lt}{data.table with variables: ihme_loc_id, year, sex, age, parent_ihme, super_region_name, region_name, empir_logit_45q15, empir_logit_5q0}
}
\value{
data.table with ihme_loc_id (of empir LT), year (of empir LT), est_ihme_loc_id, est_year, est_sim, n_match, and distance.
}
\description{
Calculate mahalanobis distance from empirical life tables to observed lifetable results, based on logit 45q15 and 5q0 values.
The Mahalanobis distance is a multi-dimensional approximation of the distance between a point and a distribution, given multiple variables to compare
Used to sub-select the most closest empirical life tables to match with observed results, across multiple variables (in this case, logit 5q0 and logit 45q15).
}
\examples{

}
