% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_lt_empirical.R
\name{download_lt_empirical}
\alias{download_lt_empirical}
\title{Download and format empirical life tables from the database}
\usage{
download_lt_empirical(run_id = "best", gbd_year)
}
\arguments{
\item{master_lts}{data.table with the variables ihme_loc_id, year, sex, source_type_id, life_table_category_id, lx* (0, 1, 5 [5] 110), empir_5q0, empir_45q15, empir_logit_5q0, empir_logit_45q15}

\item{lt_type}{character value representing the LT type to pull results for. Values:
* universal: Countries without country-specific lifetables, matching to universal lifetable database. Will use all lifetables in the universal lifetable database.
* location-specific: Countries with some country-specific non-generalizable lifetables. Will use all country-specific lifetables from the location-specific database, and then all in universal database.
* USA: United States subnational locations (but not USA national). Will only use USA subnational-specific lifetables.
* ZAF: South Africa and all subnational locations. Will only use ZAF-specific lifetables, currently coming from location-specific life_table_category_id = 2.}

\item{hiv_cdr}{data.table containing adult HIV crude death rate, to drop lifetables with high HIV results. Only used if lt_type == "large"
variables: ihme_loc_id, sex, year, adult_hiv_cdr}

\item{location}{character, ihme_loc_id for the country to pull results for}

\item{loc_map}{data.table with variables ihme_loc_id, region_id, super_region_id, parent_ihme}
}
\value{
data.table with column names:
* ihme_loc_id, year, sex, source_type
* lx(0, 1, 5 (5) 110)
* empir_5q0, empir_45q15, empir_logit_5q0, empir_logit_45q15
* region_id, super_region_id, parent_ihme
}
\description{
Format empirical life tables appropriately for processing in main lifetable process. Throw it wide by lx* and empir variables, and carry out assertions on values.
}
\examples{

}
