% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_110_mx.R
\name{gen_110_mx}
\alias{gen_110_mx}
\title{Generate mx and nLx for 110 age groups based on age 105 values}
\usage{
gen_110_mx(dt, mx_params, id_vars)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, sex, age, sim, mx, lx}

\item{mx_params}{data.table with variables: sex, parlnmx
Parameters to multiply by ln 5m105 to obtain mx at age 110 (terminal age group)}

\item{id_vars}{character vector of id variables (last one must be age)}
}
\value{
returns the given data.table with mx for the 110 age-group (also modifies the given data.table in-place?)
}
\description{
Given a data.table with lx, mx, and the 105 age group, generate mx for the 110 age group.
Applies parameters to transform the log mx at age 105 to the age 110 mx value.
}
