% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_80plus_ax.R
\name{gen_80plus_ax}
\alias{gen_80plus_ax}
\title{Calculate ax for age groups Over 80}
\usage{
gen_80plus_ax(dt, ax_params)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, sex, age, sim, qx}

\item{ax_params}{data.table with variables: sex, age, par_qx, par_sqx, par_con}

\item{ax_scale}{scale factors for ax post recalc from qx, based on adjustments needed to scale to observed 95+ mx, only applies to 95+}
}
\value{
returns the given data.table, with over-80 ax (also modifies the given data.table in-place?)
}
\description{
Calculate over-80 ax values based on a beta and constant based on qx and squared qx values
Currently no writeup of where these beta parameters come from.
}
\examples{

}
