% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hiv_scalars.R
\name{gen_hiv_scalars}
\alias{gen_hiv_scalars}
\title{Generate scalars to prevent substantial dips in HIV-free mortality due to differences between 45q15 estimation and HIV/Spectrum estimation
For all locations except for South Africa and its subnationals, ensure that the variance in the relative change of qx should be less than 0.001, and qx should never drop more than 25%, for years 1994 and onwards
For South Africa and subnationals, start measurement from 1997 onwards due to time-based discrepancies b/t HIV and 45q15, and raise variance threshold to .02 to avoid high-bump curves from making it into the analysis.}
\usage{
gen_hiv_scalars(run_id_45q15_est = "best")
}
\arguments{
\item{run_id_45q15_est}{numeric or character, either "best" for best run of 45q15 estimates or numeric for a specific 45q15 estimate run}
}
\value{
data.table of scalar values, with variables: ihme_loc_id, sex, chosen_scalar
}
\description{
Generate scalars to prevent substantial dips in HIV-free mortality due to differences between 45q15 estimation and HIV/Spectrum estimation
For all locations except for South Africa and its subnationals, ensure that the variance in the relative change of qx should be less than 0.001, and qx should never drop more than 25%, for years 1994 and onwards
For South Africa and subnationals, start measurement from 1997 onwards due to time-based discrepancies b/t HIV and 45q15, and raise variance threshold to .02 to avoid high-bump curves from making it into the analysis.
}
