% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_stage_2_life_table.R
\name{gen_stage_2_life_table}
\alias{gen_stage_2_life_table}
\title{Generate a Full Lifetable from stage-2 results, given mx, ax, and qx}
\usage{
gen_stage_2_life_table(iter_mx_qx_ax, id_vars)
}
\arguments{
\item{iter_mx_qx_ax}{data.table with variables: sim, ihme_loc_id, year, sex, age (numeric), mx, qx}

\item{id_vars}{character vector of ID variables that uniquely identify observations in the dataset. Age must be the last variable specified.}
}
\value{
data.table with variables: sim, ihme_loc_id, sex, year, age, mx, ax, qx, lx, dx, nLx, age_length, Tx, ex
}
\description{
From post-iteration stage-2 results, with qx, mx, and ax values, generate a full lifetable
}
