% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_whiv_env.R
\name{gen_whiv_env}
\alias{gen_whiv_env}
\title{Create With-HIV Envelope Results}
\usage{
gen_whiv_env(
  whiv_lt,
  population,
  u5_envelope,
  mx_params,
  hmd_indic,
  env_ids,
  gbd_year
)
}
\arguments{
\item{whiv_lt}{data.table with variables: sim, ihme_loc_id, year, sex, age (numeric), mx, lx, Tx}

\item{population}{data.table with variables: sim, ihme_loc_id, year, sex, age (numeric), population}

\item{u5_envelope}{data.table with variables: sim, ihme_loc_id, year, sex, age (character, enn/lnn/pnn/1/2/3/4), deaths}

\item{mx_params}{data.table with variables: sex, parmx}

\item{hmd_indic}{logical (T/F) -- is this country a HMD country? If not, run calc_95_no_hmd_mx}

\item{gbd_year}{to be passed into agg_u5_env function (helps select age groups to split)}
}
\value{
data.table with variables: sim, ihme_loc_id, sex, year age, deaths
        With-HIV lifetable results for all demographics, up to age 95+ (also modifies whiv_lt in-place?)
}
\description{
Create with-HIV envelope results based on with-HIV lifetables, under-5 envelope results, and population.
Deaths = mx * population, but substitute in under-5 envelope for under-5 ages. Also, modify 95+ deaths for non-HMD countries.
}
