% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_adjustment_functions.R
\name{iteration_adjustment_functions}
\alias{iteration_adjustment_functions}
\alias{stage_1_5q0_adjustment}
\alias{stage_1_45q15_adjustment}
\alias{with_hiv_5q0_adjustment}
\alias{with_hiv_45q15_adjustment}
\alias{hiv_free_ZAF_5q0_adjustment}
\alias{hiv_free_ZAF_45q15_adjustment}
\title{Functions to apply to qx iteration process.}
\usage{
stage_1_5q0_adjustment(input)

stage_1_45q15_adjustment(input)

with_hiv_5q0_adjustment(input)

with_hiv_45q15_adjustment(input)

hiv_free_ZAF_5q0_adjustment(input)

hiv_free_ZAF_45q15_adjustment(input)
}
\arguments{
\item{input}{data.table with current_scalar variable. Wide by age, either containing qx_* or mx_* variables. Stage 2 adjustments also require hiv_rr_* and abs_mx_diff.}
}
\value{
data.table with adjusted_value representing either the scaled/adjusted 5q0 or 45q15, for comparison with the target 5q0 or 45q15
}
\description{
Separate functions for stage 1 adjustment, which is same across all countries, and stage 2,
which is different for non-ZAF and ZAF (non-ZAF uses with_hiv_*_adjustment functions, while ZAF uses hiv_free_ZAF_*_adjustment).

Stage 1 5q0/45q15 adjustment: take qx_age / qx_ref_age scalar. Multiply by the scalar, then generate px_age. 
   Aggregate px values, then take 1 - agg_px = agg_qx, and use to compare to output.

Stage 2 with-HIV adjustment: adjusted_mx = mx_age + hiv_rr_age * abs_mx_diff * scalar. Getting to with-HIV mx from HIV-free mx
Where abs_mx_diff is the difference in absolute mx between the current and target mx values
Then, generate qx from ax and scaled mx, convert to px, aggregate, and generate aggregated qx

Stage 2 HIV-free ZAF adjustment: adjusted_mx = mx_age - hiv_rr_age * abs_mx_diff * scalar. Getting to HIV-free mx from with-HIV mx
Where abs_mx_diff is the difference in absolute mx between the current and target mx values
}
