% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_85plus_qx.R
\name{recalc_85plus_qx}
\alias{recalc_85plus_qx}
\title{Recalculate over-80 qx values based on HMD VR relationships with age 80 qx}
\usage{
recalc_85plus_qx(dt, betas_qx_85, id_vars)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, sex, age, sim, qx}

\item{betas_qx_85}{data.table with variables: sex, age, beta_age, beta_logitqx80, constant
Betas from logit model using selected high-quality HMD VR to convert from logit 5q80 to age-specific over-80 granular qx values
Note: age = seq(80,100,5)}

\item{id_vars}{character vector of id variables (last one must be age)}
}
\value{
data.table with variables: ihme_loc_id, sex, age, sim, qx
        Contains recalculated qx values
}
\description{
Apply high-quality HMD VR relationships between logit 5q80 and age-specific over-80 granular qx values to re-calculated over-80 granular qx
GBD2015 MortCoD Appendix: Part 3.2
Gompertz old age updates are only applied to places that don't already have those ages (e.g. aren't matched with full lifetable output)
}
