% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_ax.R
\name{recalc_ax}
\alias{recalc_ax}
\title{Recalculate ax values to produce re-optimized qx and ax values}
\usage{
recalc_ax(lifetable, id_vars)
}
\arguments{
\item{lifetable}{data.table with variables: id_vars, mx, qx, ax}

\item{id_vars}{character vector of variables that uniquely identify observations in your dataset}
}
\value{
Modifies lifetable in-place, returning modified values of ax and qx for observations that violated qx < 1 or ax >= 0 restrictions
}
\description{
For results where mx is over 1 or ax is below 0, recalculate qx directly from mx (rather than mx and ax).
 Then, recalculate ax based off of this regenerated qx value.
}
