% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_u10_nlx_mx_ax.R
\name{recalc_u10_nlx_mx_ax}
\alias{recalc_u10_nlx_mx_ax}
\title{Calculate under-10 nlx, mx, and ax values}
\usage{
recalc_u10_nlx_mx_ax(dt, id_vars)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, sex, age, sim, age_length, qx, lx, dx}

\item{id_vars}{character vector of id variables (last one must be age)}
}
\value{
data.table with variables: ihme_loc_id, sex, age, sim, age_length, qx, lx, dx
}
\description{
Calculate under-10-specific nlx, mx, and ax values based on Human Life-Table Database k1 parameters
Original k1 parameters supposed to come from Coale, A.J. and Demeny, P. 1983. Regional Model Life Tables and Stable Populations. Second Edition. Academic Press, N.Y.-L.
}
\examples{

}
