% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_hiv_cdr.R
\name{rescale_hiv_cdr}
\alias{rescale_hiv_cdr}
\title{Rescale HIV Crude Death Rates}
\usage{
rescale_hiv_cdr(
  hiv_draws,
  country_scalar,
  sex_scalar,
  hiv_ctry_exceptions,
  hiv_ctry_groups
)
}
\arguments{
\item{hiv_draws}{data.table with variables: ihme_loc_id, year, sex, sim, adult_hiv_cdr, child_hiv_cdr}

\item{country_scalar}{data.table with variables: ihme_loc_id, sex, ctry_adult_scalar, ctry_child_scalar}

\item{sex_scalar}{data.table with variables: sex, adult_scalar, child_scalar}

\item{hiv_ctry_exceptions}{data.table with variables: ihme_loc_id, sex}

\item{hiv_ctry_groups}{data.table with variables: ihme_loc_id, sex, group (character)}
}
\value{
data.table with variables: ihme_loc_id, year, sex, sim, adult_hiv_cdr, child_hiv_cdr
}
\description{
Use country/sex and sex-specific scalars to scale child and adult HIV CDR
}
