% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scalars.R
\name{select_scalars}
\alias{select_scalars}
\title{Run the iteration process to select the best scalars for the stage 1 and 2 iterations.}
\usage{
select_scalars(
  input,
  hiv_type,
  num_iterations,
  num_scalars,
  target_var,
  adjustment_function
)
}
\arguments{
\item{input}{data.table to be processed. Contains all variables needed in the given adjustment_function, along with min_scalar and max_scalar variables.
If hiv_type == "hiv_free_ZAF", must also include scalar_cap variable.}

\item{hiv_type}{character, name of HIV processing type used for scalar iteration. Used to determine whether to apply caps on scalar values}

\item{num_iterations}{numeric, the number of time select_scalars should create a new set of scalars and find an optimal scalar value within the set}

\item{num_scalars}{numeric, the number of values to generate and test within each scalar_set}

\item{target_var}{character, name of variable that is being compared to (target variable)}

\item{adjustment_function}{numeric whole number, the number of scalar values to be tested within each iteration}
}
\value{
numeric value representing the optimal scalar given the input and other argument definitions
}
\description{
Run scalar selection process within the stage 1 and 2 iteration processes. 
Test x equally-spaced scalar values. After finding the closest fit to target value,
re-center on the best scalar value and redefine the scalar value window to 4 times the step-size of the current-run (narrowing the window).
Then, run another iteration on new scalar values. Repeat for num_iterations times.
}
