% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan_lts.R
\name{smooth_stan_lts}
\alias{smooth_stan_lts}
\title{Smooth post-standard LT matching lifetables for selected countries}
\usage{
smooth_stan_lts(stan_qx, id_vars)
}
\arguments{
\item{id_vars}{character vector of id variables (should not include year), E.g. c("ihme_loc_id", "sex", "age", "sim")}

\item{dt}{data.table with variables: ihme_loc_id, year, sex, age, sim, stan_qx}
}
\value{
data.table with variables: ihme_loc_id, year, sex, age, sim, stan_qx
        Contains smoothed qx values, should be same nrows as dt
}
\description{
To adjust for high-variance standard life table results over time, smooth the output qx values over years (but within id_vars such as age, country, sex, etc.)
}
\examples{
test <- CJ(ihme_loc_id = "USA", year = c(1950:2016), sex = c("male", "female"), age = c(0, 1, seq(5, 5, 110)), sim = 1:3)
test[, qx := (age/110) + 1950/year-.5]
results <- smooth_stan_lts(test, id_vars = c("ihme_loc_id", "sex", "age", "sim"))

}
