% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_hiv_mx.R
\name{subtract_hiv_mx}
\alias{subtract_hiv_mx}
\title{Subtract HIV cause-specific mx values from 45q15 and 5q0}
\usage{
subtract_hiv_mx(qx_draws, hiv_draws, hiv_prop_range, hiv_ctry_groups)
}
\arguments{
\item{qx_draws}{data.table with variables: ihme_loc_id, year, sex, sim, entry_45q15, entry_5q0
With-HIV results from the 5q0 and 45q15 processes (location/age/sex/year/draw-specific)}

\item{hiv_draws}{data.table with variables: ihme_loc_id, year, sex, adult_hiv_cdr, child_hiv_cdr. Post-rescaling HIV CDR based on various scaling values. Now mean values rather than draw-specific.}

\item{hiv_prop_range}{data.table with variables: ihme_loc_id, sex, upper_adult, upper_child. Generated by Spectrum, perhaps the upper 97.5 percentile bound on HIV deaths/envelope proportion}

\item{hiv_ctry_groups}{data.table with variables: ihme_loc_id, sex, group (character). Used to generate a floor for HIV-free values in high-HIV countries that risk having a false floor of HIV.}
}
\value{
data.table with variables: ihme_loc_id, year, sex, sim, calc_45q15, calc_5q0
}
\description{
Subtract HIV cause-specific mx, bounded by HIV proportion proportion values, from 45q15 and 5q0 to generate HIV-free 45q15 and 5q0 values
}
